Moves EphemeralErrorHandler to logs/ephemeral
