+++ assertions/BUILD.bazel
@@ -23,7 +23,7 @@ go_library(
         "//runtime",
         "//solgen/go/rollupgen",
         "//util/stopwaiter",
-        "//util/log",
+        "//logs/ephemeral",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind",
         "@com_github_ethereum_go_ethereum//common",
         "@com_github_ethereum_go_ethereum//core/types",
