+++ assertions/confirmation.go
@@ -12,8 +12,8 @@ import (
 	solimpl "github.com/offchainlabs/bold/chain-abstraction/sol-implementation"
 	"github.com/offchainlabs/bold/challenge-manager/types"
 	"github.com/offchainlabs/bold/containers/option"
+	"github.com/offchainlabs/bold/logs/ephemeral"
 	retry "github.com/offchainlabs/bold/runtime"
-	utillog "github.com/offchainlabs/bold/util/log"
 )
 
 func (m *Manager) queueCanonicalAssertionsForConfirmation(ctx context.Context) {
@@ -58,7 +58,7 @@ func (m *Manager) keepTryingAssertionConfirmation(ctx context.Context, assertion
 	}
 
 	backoffLogLevel := time.Second
-	exceedsMaxMempoolSizeEphemeralErrorHandler := utillog.NewEphemeralErrorHandler(10*time.Minute, "posting this transaction will exceed max mempool size", 0)
+	exceedsMaxMempoolSizeEphemeralErrorHandler := ephemeral.NewEphemeralErrorHandler(10*time.Minute, "posting this transaction will exceed max mempool size", 0)
 
 	ticker := time.NewTicker(m.confirmationAttemptInterval)
 	defer ticker.Stop()
