+++ challenge-manager/chain-watcher/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//runtime",
         "//solgen/go/challengeV2gen",
         "//util/stopwaiter",
-        "//util/log",
+        "//logs/ephemeral",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind",
         "@com_github_ethereum_go_ethereum//common",
         "@com_github_ethereum_go_ethereum//log",
