+++ logs/ephemeral/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "log",
+    name = "ephemeral",
     srcs = ["log.go"],
-    importpath = "github.com/OffchainLabs/bold/util/log",
+    importpath = "github.com/OffchainLabs/bold/logs/ephemeral",
     visibility = ["//visibility:public"],
     deps = ["@com_github_ethereum_go_ethereum//log"],
 )
