+++ logs/ephemeral/log_test.go
@@ -1,4 +1,4 @@
-package log
+package ephemeral
 
 import (
 	"errors"
